% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionDB.R
\name{OhdsiDataCatalogue}
\alias{OhdsiDataCatalogue}
\title{OhdsiDataCatalogue Function}
\usage{
OhdsiDataCatalogue(Server, IP, Schema, ID, PW)
}
\arguments{
\item{Server, IP, Schema, ID, PW}{connection details.}
}
\description{
This function allows you to see basic elements of the data catalogue automatically.
}
\examples{
OhdsiDataCatalogue("sql server","IP","cdmdatabaseSchema","ID","PW")
}
\keyword{catalogue}
\keyword{data}
